[refactor-buy/cart]: 컴포넌트 분리
cart페이지 컴포넌트 분리
 items: 장바구니 아이탬이 출력되는 그리드

nonInCart: 장바구니 아이탬이 없을때 나오는 ui

processNav: 주문완료까지 프로세스 현황바

totalInfo: 선택아이템에 대한 최종 결제금액 및 배송비

*현재 items의 데이터 타입을 보면 상품옵션에 대해 size,color,other만 호환가능
이를 타입을 분리해서 장바구니 데이터 저장시 option이라는 키값을 만들고 내부 타입을
[key:string]:string의 객체 타입으로 변경계획 중

관련노션: https://www.notion.so/56e674cbb5234b3eb8ab182ac7c15c40?v=41f8967bcafe4f3a83ff07ae9417a5ce&p=d6da261acc304b52b70473dc28b59fa6&pm=s
