+++ services/buy/src/App.tsx
@@ -1,7 +1,7 @@
-import React, { Suspense, useEffect } from "react";
+import React, { Suspense } from "react";
 import { Routes, Route } from "react-router-dom";
 import Globalstyles from "../../../shared/shared/globalStyle/GlobalStyle";
-import CartPage from "./page/CartPage";
+import CartPage from "./page/cart/CartPage";
 import Layout from "../../../shared/shared/layout/Layout";
 
 function App() {
