+++ services/buy/src/api/Category.ts
@@ -1,15 +1,13 @@
 import { BASE_URL } from "./const";
 import axios from "axios";
 
-export const getCategoryData = async (path:string,page:number) => {
-    try {
-      const res = await axios.get(`${BASE_URL}/category/${path}?page=${page}`);
-        console.log(res)
+export const getCategoryData = async (path: string, page: number) => {
+  try {
+    const res = await axios.get(`${BASE_URL}/category/${path}?page=${page}`);
 
-      return res
-    } catch (error) {
-      console.error("Error fetching main data:", error);
-      throw error; 
-    }
-  };
-
+    return res;
+  } catch (error) {
+    console.error("Error fetching main data:", error);
+    throw error;
+  }
+};
